Update app with new screens and widgets
