+++ app/lib/main.dart
@@ -22,7 +22,7 @@ import 'package:spendwise/screens/login_screen.dart';
 import 'package:spendwise/screens/main_screen.dart';
 import 'package:spendwise/screens/register_screen_screen.dart';
 import 'package:spendwise/screens/edit_data_screens/select_%20monetary_unit.dart';
-import 'package:spendwise/screens/startup_screen.dart';
+import 'package:spendwise/screens/static_screens/startup_screen.dart';
 
 // utils
 import 'package:spendwise/utils/fetch_all_data.dart';
