+++ app/lib/screens/pages/home_page.dart
@@ -1,22 +1,29 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
+
+// Package imports:
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:spendwise/data/constant_values.dart';
-import 'package:spendwise/model/transaction.dart';
-import 'package:spendwise/provider/monetary_units.dart';
+import 'package:spendwise/widgits/cards/balance_card.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
+//  Model:
+import 'package:spendwise/model/transaction.dart';
+
+//Provider:
 import 'package:spendwise/provider/transaction_provider.dart';
 import 'package:spendwise/provider/user_provider.dart';
-import 'package:spendwise/screens/edit_data_screens/add_update_transaction_screen.dart';
-import 'package:spendwise/screens/statics_data_screens/setting_screen.dart';
-import 'package:spendwise/screens/statics_data_screens/user_detail_screen.dart';
-import 'package:spendwise/screens/view_all_transaction_screen.dart';
-import 'package:spendwise/utils/fetch_all_data.dart';
 
-import 'package:spendwise/widgits/cards/transaction_card.dart';
-import 'package:visibility_detector/visibility_detector.dart';
-import 'package:shared_preferences/shared_preferences.dart';
+// Screens & Widgets
+import 'package:spendwise/screens/edit_data_screens/add_update_transaction_screen.dart';
+import 'package:spendwise/screens/static_screens/setting_screen.dart';
+import 'package:spendwise/screens/static_screens/user_detail_screen.dart';
+
+//  Utils:
+import 'package:spendwise/utils/fetch_all_data.dart';
+import 'package:spendwise/widgits/past_transaction.dart';
 
 enum TransactionFilter {
   byMonth,
@@ -51,37 +58,41 @@ class HomeScreen extends HookConsumerWidget {
 
     return Scaffold(
       appBar: AppBar(
-          title: const Text(
-            "Home",
+        title: const Text("Home"),
+        actions: [
+          // User Detail Button
+          IconButton(
+            onPressed: () {
+              Navigator.push(context, MaterialPageRoute(builder: (context) {
+                return const UserDetailScreen();
+              }));
+            },
+            icon: Icon(
+              MdiIcons.accountCircleOutline,
+              color: Theme.of(context).colorScheme.onBackground,
+              size: 25,
+            ),
           ),
-          actions: [
-            IconButton(
+
+          // Setting Button
+          IconButton(
               onPressed: () {
                 Navigator.push(context, MaterialPageRoute(builder: (context) {
-                  return const UserDetailScreen();
+                  return SettingScreen();
                 }));
               },
               icon: Icon(
-                MdiIcons.accountCircleOutline,
+                MdiIcons.cog,
                 color: Theme.of(context).colorScheme.onBackground,
-                size: 25,
-              ),
-            ),
-            IconButton(
-                onPressed: () {
-                  Navigator.push(context, MaterialPageRoute(builder: (context) {
-                    return SettingScreen();
-                  }));
-                },
-                icon: Icon(
-                  MdiIcons.cog,
-                  color: Theme.of(context).colorScheme.onBackground,
-                ))
-          ]),
+              ))
+        ],
+      ),
       body: RefreshIndicator(
+        // refresh indicator
         onRefresh: () async {
           await fetchData(ref);
         },
+
         child: Padding(
             padding: const EdgeInsets.only(
               left: padding,
@@ -93,28 +104,30 @@ class HomeScreen extends HookConsumerWidget {
                   child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
-                  // const SizedBox(
-                  //   height: 40,
-                  // ),
-                  header(context, ref),
+                  _header(context, ref),
                   const SizedBox(
                     height: 40,
                   ),
-                  balanceCard(context, ref, transactions, isVisibile),
+                  BalanceCards(
+                    transactions: transactions,
+                    isVisible: isVisibile,
+                  ),
                   const SizedBox(
                     height: 20,
                   ),
-                  filterChips(
+                  _catergoryChips(
                     context,
                   ),
                   const SizedBox(
                     height: 40,
                   ),
-                  pastTransactions(context, transactions, filter, ref),
+                  PastTransactions(transactions: transactions, filter: filter)
                 ],
               )),
             )),
       ),
+
+      // Floating Action Button for adding new transaction if the user is above button is not visible
       floatingActionButton: isVisibile.value
           ? FloatingActionButton.large(
               backgroundColor: Theme.of(context).colorScheme.tertiaryContainer,
@@ -133,7 +146,7 @@ class HomeScreen extends HookConsumerWidget {
     );
   }
 
-  Widget header(
+  Widget _header(
     BuildContext context,
     WidgetRef ref,
   ) {
@@ -162,133 +175,7 @@ class HomeScreen extends HookConsumerWidget {
     );
   }
 
-  Widget balanceCard(
-    BuildContext context,
-    WidgetRef ref,
-    List<Transaction> transactions,
-    ValueNotifier<bool> isVisible,
-  ) {
-    double getTotalBalance() {
-      double total = 0;
-      for (var account in ref.watch(userProvider).accounts) {
-        total += account.balance;
-      }
-      return total;
-    }
-
-    return SizedBox(
-      height: 200,
-      child: IntrinsicHeight(
-        child: Row(
-          children: [
-            Container(
-              width: (MediaQuery.of(context).size.width - 20) * .6,
-              decoration: BoxDecoration(
-                color: Theme.of(context)
-                    .colorScheme
-                    .primaryContainer
-                    .withOpacity(.8),
-                borderRadius: BorderRadius.circular(30),
-              ),
-              padding: const EdgeInsets.all(20),
-              child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      "Total Balance",
-                      style: Theme.of(context).textTheme.labelLarge!.copyWith(
-                            color: Theme.of(context).colorScheme.onBackground,
-                            fontWeight: FontWeight.bold,
-                          ),
-                    ),
-                    const SizedBox(
-                      height: 10,
-                    ),
-                    FittedBox(
-                      child: Text(
-                        "${ref.watch(monetaryUnitProvider.notifier).get()} ${getTotalBalance().toStringAsFixed(2)}",
-                        style: Theme.of(context)
-                            .textTheme
-                            .displayMedium!
-                            .copyWith(
-                              color: Theme.of(context).colorScheme.onBackground,
-                            ),
-                      ),
-                    ),
-                    const SizedBox(
-                      height: 10,
-                    ),
-                    Text(
-                      "Total Expenses of Month",
-                      style: Theme.of(context).textTheme.labelLarge!.copyWith(
-                            color: Theme.of(context).colorScheme.onBackground,
-                            fontWeight: FontWeight.bold,
-                          ),
-                    ),
-                    const SizedBox(
-                      height: 5,
-                    ),
-                    FittedBox(
-                      child: Text(
-                        "${ref.watch(monetaryUnitProvider.notifier).get()}${ref.watch(transactionProvider.notifier).totalExpensesByMonth(DateTime.now().month, DateTime.now().year).abs().toStringAsFixed(2)}",
-                        style: Theme.of(context)
-                            .textTheme
-                            .headlineMedium!
-                            .copyWith(
-                              color: Theme.of(context).colorScheme.onBackground,
-                            ),
-                      ),
-                    ),
-                  ]),
-            ),
-            const Spacer(),
-            VisibilityDetector(
-              key: const Key("add_transaction_button"),
-              onVisibilityChanged: (visibilityInfo) {
-                double visiblePercentage = visibilityInfo.visibleFraction * 100;
-                if (visiblePercentage > 60) {
-                  isVisible.value = false;
-                } else {
-                  isVisible.value = true;
-                }
-              },
-              child: Material(
-                color: Theme.of(context)
-                    .colorScheme
-                    .tertiaryContainer
-                    .withOpacity(.8),
-                borderRadius: BorderRadius.circular(30),
-                child: InkWell(
-                  onTap: () {
-                    Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => AddTransactionScreen()));
-                  },
-                  borderRadius: BorderRadius.circular(30),
-                  splashColor: Theme.of(context).colorScheme.tertiaryContainer,
-                  highlightColor:
-                      Theme.of(context).colorScheme.tertiaryContainer,
-                  child: Container(
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(30),
-                    ),
-                    width: (MediaQuery.of(context).size.width - 20) * .35,
-                    height: double.infinity,
-                    child: Icon(MdiIcons.plus,
-                        color: Theme.of(context).colorScheme.onBackground,
-                        size: 50),
-                  ),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget filterChips(BuildContext context) {
+  Widget _catergoryChips(BuildContext context) {
     return SizedBox(
       width: double.infinity,
       child: Row(
@@ -296,191 +183,40 @@ class HomeScreen extends HookConsumerWidget {
         children: [
           ...TransactionCatergory.values
               .sublist(0, 5)
-              .map((e) => Ink(
-                    decoration: BoxDecoration(
-                      color: Theme.of(context)
-                          .colorScheme
-                          .primaryContainer
-                          .withOpacity(.8),
-                      borderRadius: BorderRadius.circular(10),
-                    ),
-                    child: InkWell(
-                      onTap: () {
-                        Navigator.push(
-                            context,
-                            MaterialPageRoute(
-                                builder: (context) => AddTransactionScreen(
-                                      userSelectedCategory: e,
-                                    )));
-                      },
-                      borderRadius: BorderRadius.circular(10),
-                      splashColor:
-                          Theme.of(context).colorScheme.primaryContainer,
-                      child: Padding(
-                        padding: const EdgeInsets.all(8.0),
-                        child: Icon(
-                          getTransactionCatergoryIcon(e),
-                          color: Theme.of(context).colorScheme.primary,
-                          size: 30,
-                        ),
+              .map(
+                (e) => Ink(
+                  decoration: BoxDecoration(
+                    color: Theme.of(context)
+                        .colorScheme
+                        .primaryContainer
+                        .withOpacity(.8),
+                    borderRadius: BorderRadius.circular(10),
+                  ),
+                  child: InkWell(
+                    onTap: () {
+                      Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                              builder: (context) => AddTransactionScreen(
+                                    userSelectedCategory: e,
+                                  )));
+                    },
+                    borderRadius: BorderRadius.circular(10),
+                    splashColor: Theme.of(context).colorScheme.primaryContainer,
+                    child: Padding(
+                      padding: const EdgeInsets.all(8.0),
+                      child: Icon(
+                        getTransactionCatergoryIcon(e),
+                        color: Theme.of(context).colorScheme.primary,
+                        size: 30,
                       ),
                     ),
-                  ))
+                  ),
+                ),
+              )
               .toList()
         ],
       ),
-    );
-  }
-
-  Widget pastTransactions(
-    BuildContext context,
-    List<Transaction> transactions,
-    ValueNotifier<TransactionFilter> filter,
-    WidgetRef ref,
-  ) {
-    void setFilter() {
-      showModalBottomSheet(
-          context: context,
-          builder: (context) {
-            return Container(
-              padding: const EdgeInsets.all(10),
-              height: 160,
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  ListTile(
-                    title: Text("Transactions of Month",
-                        style: Theme.of(context).textTheme.bodyLarge!.copyWith(
-                              color: Theme.of(context).colorScheme.onBackground,
-                              fontWeight: FontWeight.bold,
-                            )),
-                    onTap: () async {
-                      filter.value = TransactionFilter.byMonth;
-                      final SharedPreferences prefs = await _prefs;
-                      prefs.setString("filter", "month");
-                      if (context.mounted) {
-                        Navigator.pop(context);
-                      }
-                    },
-                  ),
-                  ListTile(
-                    title: Text(
-                      "Transactions of Week",
-                      style: Theme.of(context).textTheme.bodyLarge!.copyWith(
-                            color: Theme.of(context).colorScheme.onBackground,
-                            fontWeight: FontWeight.bold,
-                          ),
-                    ),
-                    onTap: () async {
-                      filter.value = TransactionFilter.byWeek;
-                      final SharedPreferences prefs = await _prefs;
-                      prefs.setString("filter", "week");
-                      if (context.mounted) {
-                        Navigator.pop(context);
-                      }
-                    },
-                  ),
-                ],
-              ),
-            );
-          });
-    }
-
-    double getTotalExpenses() {
-      double total = 0;
-      for (var transaction in transactions) {
-        if (transaction.type == TransactionType.expense) {
-          total -= transaction.amount;
-        } else {
-          total += transaction.amount;
-        }
-      }
-
-      return total;
-    }
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: [
-        SizedBox(
-          width: double.infinity,
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              InkWell(
-                borderRadius: BorderRadius.circular(10),
-                onTap: () {
-                  setFilter();
-                },
-                child: Padding(
-                  padding: const EdgeInsets.all(5),
-                  child: Row(
-                    children: [
-                      Text(
-                        filter.value == TransactionFilter.byMonth
-                            ? "All Transactions of Month"
-                            : "All Transactions of Week",
-                        textAlign: TextAlign.left,
-                        style: Theme.of(context).textTheme.bodyLarge!.copyWith(
-                            color: Theme.of(context).colorScheme.onBackground,
-                            fontWeight: FontWeight.bold),
-                      ),
-                      Icon(
-                        MdiIcons.unfoldMoreHorizontal,
-                        color: Theme.of(context).colorScheme.onBackground,
-                        size: 20,
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-              Text(
-                "${ref.watch(monetaryUnitProvider.notifier).get()}${getTotalExpenses().abs().toStringAsFixed(2)}",
-                textAlign: TextAlign.right,
-                style: Theme.of(context).textTheme.bodyLarge!.copyWith(
-                      color: getTotalExpenses() >= 0
-                          ? Colors.green
-                          : Theme.of(context).colorScheme.error,
-                      fontWeight: FontWeight.bold,
-                    ),
-              ),
-            ],
-          ),
-        ),
-        const SizedBox(
-          height: 20,
-        ),
-        ...transactions.map((e) => TransactionCard(transaction: e)).toList(),
-        transactions.isNotEmpty
-            ? Container(
-                margin: const EdgeInsets.only(top: 20),
-                child: FilledButton.tonal(
-                    onPressed: () {
-                      Navigator.push(context,
-                          MaterialPageRoute(builder: (context) {
-                        return ViewAllTransactionScreen(
-                          title: "All Transactions",
-                          transactions: ref
-                              .read(transactionProvider.notifier)
-                              .getSorted(),
-                        );
-                      }));
-                    },
-                    child: Text(
-                      "View All",
-                      softWrap: false,
-                      style: Theme.of(context).textTheme.labelMedium!.copyWith(
-                            fontWeight: FontWeight.bold,
-                          ),
-                    )))
-            : Text(
-                "No Transactions",
-                style: Theme.of(context).textTheme.bodyLarge!.copyWith(),
-              ),
-        const SizedBox(
-          height: 80,
-        )
-      ],
     );
   }
 }
