+++ app/lib/screens/static_screens/startup_screen.dart
@@ -1,6 +1,11 @@
+// Flutter imports
 import 'package:flutter/material.dart';
+
+// Package imports
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
+
+// Screens & Widgets
 import 'package:spendwise/data/constant_values.dart';
 import 'package:spendwise/utils/about_device.dart';
